Gtag Removed and stored in env
