+++ src/pages/landingPage.jsx
@@ -7,7 +7,7 @@ import SelectButton from '../components/buttons/select_button.jsx';
 import landingPageBg from '../assets/landingpagebg.svg';
 import Navbar from '../components/navbar.jsx';
 import getStates from "../actions/getStates";
-import { trackPageView, trackEvent } from "../services/analytics";
+import { trackPageView, trackEvent, initializeAnalytics } from "../services/analytics";
 
 const LandingPage = () => {
     const navigate = useNavigate();
@@ -43,28 +43,14 @@ const LandingPage = () => {
     };
 
     useEffect(() => {
+      // Initialize Google Analytics
+      initializeAnalytics();
+      
       // Track page view when component mounts
       trackPageView('/');
       
       // Load states data
       getStatesData();
-      
-      // Add Google Analytics script if not already present
-      if (!window.gtag) {
-        const script1 = document.createElement("script");
-        script1.async = true;
-        script1.src = "https://www.googletagmanager.com/gtag/js?id=G-Q885CFPDYX";
-        document.head.appendChild(script1);
-
-        const script2 = document.createElement("script");
-        script2.innerHTML = `
-          window.dataLayer = window.dataLayer || [];
-          function gtag(){dataLayer.push(arguments);}
-          gtag('js', new Date());
-          gtag('config', 'G-Q885CFPDYX');
-        `;
-        document.head.appendChild(script2);
-      }
     }, []);
     
     return (
